app/pipeline/inference/modules/sceneUnderstandingModule/requirements.txt
clarifai_grpc==8.8.0
pika==1.3.2
numpy==1.22.4
python-decouple==3.8
boto3==1.26.118
requests==2.31.0